AudioContext still doesn't entirely reset JavaSound when stopped and started again.
UGen now tests for connected input UGen. 

Added TimeGraph - visualisation of clock for GUI elements.

Changed Environment so that environment is created with a given AudioContext.

Modified Player so that song groups can be flipped on quantised time.

Modified SongGrid so that cntrl-click creates popup to choose channels and pathways.