A few Sample related thingymajigs.
Recorder has parameters that affect its sample resizing behaviour.

Sample.regionMaster is now in control of all region loading. It is a ThreadPoolExecutor and so has a bunch of threads that it uses to load regions in sample.

Also build.xml has been updated to build a processing version of beads that has just requires "import beads.*;" to import everything.