Major attempt to refactor AudioContext.
AudioContext is now independent of the IO implementation, so it can run offline without using JavaSound, or use a different IO option. The class AudioIO now exists, which is the abstract base class for creating IO implementations. The standard JavaSound version lives in the package net.beadsrproject.beads.core.io and is called JavaSoundAudioIO. This is still created as the default. 

The only breakage is on the constructors for AudioContext which referred to the system buffer size (as opposed to the Beads internal buffer size). These have been replaced with constructors which take an AudioIO class as an argument. 

In this redesign, AudioContext will also handle setting up audio inputs (as well as outputs) by forwarding the request to the AudioIO. This uses the new method getAudioInput(int[]), which returns a UGen providing audio inputs for the channels specified in the int array.