At last Ollie sorts out JavaSound audio input properly. Clean recording.
Also redesigned the checkForDroppedFrames behaviour. All that happens now is the time of the update itself is calculated and checked against the duration of an audio frame. This is not such a literal check for whether things have taken as long as they should, and it can be misleading since all kinds of other processes might be using cpu, but it's a lot cleaner.