Added JackControl to the jjack io subproject. Use this to run jackdmp. Currently not configurable and set to coreaudio.
More work on Analyzer: added methods to listen to beats. As with the analysis packages, this is messy stuff and should be reworked soon.

Added OnsetView and FloatArrayPlot to beads_player.gui package. Made some mods to LevelMeter. Again, these player classses are experimental and need reworking. In particular, what is a good strategy for calling repaint on components? e.g., LevelMeter currently runs its own thread which is very silly.